+++ pints/residuals_diagnostics.py
@@ -5,7 +5,6 @@
 # released under the BSD 3-clause license. See accompanying LICENSE.md for
 # copyright notice and full license details.
 #
-from __future__ import absolute_import, division
 import math
 import numpy as np
 import scipy.special