+++ pints/tests/test_io.py
@@ -6,7 +6,6 @@
 # released under the BSD 3-clause license. See accompanying LICENSE.md for
 # copyright notice and full license details.
 #
-from __future__ import print_function, unicode_literals
 import os
 import pints
 import pints.io
@@ -14,12 +13,6 @@ import numpy as np
 import unittest
 
 from shared import TemporaryDirectory
-
-# Consistent unit testing in Python 2 and 3
-try:
-    unittest.TestCase.assertRaisesRegex
-except AttributeError:
-    unittest.TestCase.assertRaisesRegex = unittest.TestCase.assertRaisesRegexp
 
 
 class TestIO(unittest.TestCase):