+++ pints/tests/test_mcmc_haario_bardenet_ac.py
@@ -14,12 +14,6 @@ import numpy as np
 
 from shared import StreamCapture
 
-# Consistent unit testing in Python 2 and 3
-try:
-    unittest.TestCase.assertRaisesRegex
-except AttributeError:
-    unittest.TestCase.assertRaisesRegex = unittest.TestCase.assertRaisesRegexp
-
 
 class TestHaarioBardenetACMC(unittest.TestCase):
     """