+++ pints/tests/test_nested_ellipsoid_sampler.py
@@ -12,12 +12,6 @@ import numpy as np
 import pints
 import pints.toy
 
-# Unit testing in Python 2 and 3
-try:
-    unittest.TestCase.assertRaisesRegex
-except AttributeError:
-    unittest.TestCase.assertRaisesRegex = unittest.TestCase.assertRaisesRegexp
-
 
 class TestNestedEllipsoidSampler(unittest.TestCase):
     """