+++ pints/tests/test_opt_gradient_descent.py
@@ -18,12 +18,6 @@ from shared import CircularBoundaries
 debug = False
 method = pints.GradientDescent
 
-# Consistent unit testing in Python 2 and 3
-try:
-    unittest.TestCase.assertRaisesRegex
-except AttributeError:
-    unittest.TestCase.assertRaisesRegex = unittest.TestCase.assertRaisesRegexp
-
 
 class TestGradientDescent(unittest.TestCase):
     """