+++ pints/tests/test_opt_pso.py
@@ -15,11 +15,6 @@ import pints.toy
 
 from shared import StreamCapture, TemporaryDirectory, CircularBoundaries
 
-# Consistent unit testing in Python 2 and 3
-try:
-    unittest.TestCase.assertRaisesRegex
-except AttributeError:
-    unittest.TestCase.assertRaisesRegex = unittest.TestCase.assertRaisesRegexp
 
 debug = False
 method = pints.PSO