+++ pints/toy/_german_credit.py
@@ -5,17 +5,10 @@
 # released under the BSD 3-clause license. See accompanying LICENSE.md for
 # copyright notice and full license details.
 #
-from __future__ import absolute_import, division
-from __future__ import print_function, unicode_literals
 import numpy as np
 import io
 from scipy import stats
 from . import ToyLogPDF
-
-try:
-    from urllib.request import urlopen
-except ImportError:     # pragma: no python 3 cover
-    from urllib2 import urlopen
 
 
 class GermanCreditLogPDF(ToyLogPDF):